trial_utilities
* dprime calculation
